requirements.txt
django
djangorestframework
djangorestframework-simplejwt
drf-yasg
pillow
django-cors-headers
celery
redis
python-decouple
psycopg2-binary
django-cors-headers